feat(map): refactor and enhance map components with clustering and shared utilities
- Introduced clustering functionality for photo markers, improving the display of grouped photos on the map.
- Added shared components for GeoJSON layers, cluster markers, and photo marker pins to streamline code and enhance reusability.
- Updated Mapbox and MapLibre components to utilize new shared utilities, improving maintainability and consistency across the application.
- Enhanced styling and interactions for cluster markers and photo markers, providing a better user experience.
