+++ apps/web/src/components/ui/map/index.ts
@@ -3,3 +3,4 @@ export * from './MapBackButton'
 export * from './Mapbox'
 export * from './MapInfoPanel'
 export * from './MapLoadingState'
+export * from './shared'
